Fix @gautierronan comments
