+++ torchqdynamics/__init__.py
@@ -1,5 +1,5 @@
 from .utils import *
 from .mesolve import mesolve
-from .mesolve import solver_options as mesolver
+from .solver import *
 from .ssolve import ssolve
 from .smesolve import smesolve
